podcast_llm/requirements.txt
langchain==0.3.27
langchain-community==0.3.14
langchain-core==0.3.21
langchain-openai==0.2.14
langchain-google-genai==2.0.9
langchain-anthropic==0.2.11
tavily-python==0.5.0
PyYAML==6.0.2
python-dotenv==1.0.1
pydantic==2.9.2
gradio==5.10.0
newspaper3k==0.2.8
PyPDF2==3.0.1
python-docx==1.1.2
youtube-transcript-api==0.6.2
openai==1.59.1
google-generativeai==0.8.3
anthropic==0.39.0
python-docx==4.12.2
#playsound==1.3.0
simpleaudio==.0.4
pydub==0.25.1